refactors folder structure
