+++ src/routes/events/components/events/_EventsHero.svelte
