+++ src/routes/events/components/online/_ActivityType.svelte
