+++ src/routes/events/components/online/_ActivityTypes.svelte
